requirements.txt
Flask==3.1.0
mysql-connector-python==9.1.0
bcrypt==4.2.1
qrcode[pil]==8.0
reportlab==4.2.5
Pillow>=11.1.0
python-dotenv==1.0.1
gunicorn==22.0.0
